refactor: organise constrained sampling by responsibility
