+++ cicd/tests/test_multi_ellipsoid.py
@@ -8,7 +8,7 @@ from jaxctx.priors.prior import Prior
 
 from jaxns.mixed_precision import mp_policy
 from jaxns.model import Model
-from jaxns.multi_ellipsoid_utils import (
+from jaxns.sampling.ellipsoid import (
     EllipsoidParams,
     bounding_ellipsoid,
     circle_to_ellipsoid,
