+++ src/jaxns/constrained_sampler.py
@@ -3,23 +3,54 @@ from __future__ import annotations
 import dataclasses
 import warnings
 from abc import ABC, abstractmethod
-from collections.abc import Callable
 from functools import partial
 from typing import Any, NamedTuple
 
 import jax
-import numpy as np
 from jax import numpy as jnp
 from jax import random
 
 from jaxns.cumulative_ops import cumulative_op_static
 from jaxns.mixed_precision import mp_policy
 from jaxns.model import Model
-from jaxns.multi_ellipsoid_utils import SamplerData, empty_sampler_data
-from jaxns.pytree import PureDataclassPytree, TreeField, pytree_ravel
-from jaxns.random_utils import sample_uniformly_masked
-from jaxns.samples import PhantomSamples, Samples, SeedPoint
-from jaxns.types import BoolArray, FloatArray, IntArray, PRNGKey, UType
+from jaxns.pytree import PureDataclassPytree, TreeField
+from jaxns.samples import PhantomSamples, SeedPoint
+from jaxns.sampling.batching import (
+    evaluate_request,
+)
+from jaxns.sampling.batching import (
+    sample_complete_chains as _sample_complete_chains,
+)
+from jaxns.sampling.continuation import (
+    _continue_slice_chains,
+)
+from jaxns.sampling.ellipsoid import SamplerData, empty_sampler_data
+from jaxns.sampling.protocol import (
+    ConstrainedSampleBatch,
+    ConstrainedSampleRequest,
+    LikelihoodEvaluation,
+    LikelihoodRequest,
+)
+from jaxns.sampling.seeding import get_seed_point
+from jaxns.sampling.slice import (
+    _draw_ellipsoidal_direction,
+    _new_proposal,
+    _sample_direction,
+)
+from jaxns.types import BoolArray, FloatArray, IntArray, UType
+
+__all__ = [
+    "AbstractSampler",
+    "ConstrainedSampleBatch",
+    "ConstrainedSampleRequest",
+    "EllipsoidalDirection",
+    "LikelihoodEvaluation",
+    "LikelihoodRequest",
+    "UniDimSliceSampler",
+    "evaluate_request",
+    "get_seed_point",
+    "sample_request",
+]
 
 # Continuation bookkeeping dominates cheap, short or narrow batches. The
 # issue-244 matrix shows material physical-work savings at 32 slice transitions
@@ -29,87 +60,6 @@ from jaxns.types import BoolArray, FloatArray, IntArray, PRNGKey, UType
 MIN_CONTINUATION_SLICES = 32
 MIN_CONTINUATION_CHAINS = 8
 
-
-@dataclasses.dataclass(slots=True, frozen=True)
-class LikelihoodRequest(PureDataclassPytree):
-    """Unit-hypercube points whose likelihoods must run on a worker."""
-
-    U_samples: UType  # [S, ...] unit-hypercube pytree leaves
-
-
-LikelihoodRequest.register_pytree()
-
-
-@dataclasses.dataclass(slots=True, frozen=True)
-class LikelihoodEvaluation(PureDataclassPytree):
-    """Worker-computed likelihood values aligned with one request."""
-
-    log_likelihoods: FloatArray  # [S]
-
-
-LikelihoodEvaluation.register_pytree()
-
-
-def evaluate_request(
-        model: Model,
-        request: LikelihoodRequest,
-        *,
-        args=(),
-        params=None,
-) -> LikelihoodEvaluation:
-    """Evaluate likelihoods without involving constrained-chain state."""
-
-    def evaluate_one(U):
-        return model.log_likelihood(
-            U,
-            args=args,
-            params=params,
-            allow_nan=False,
-        ).astype(mp_policy.measure_dtype)
-
-    batch_size = jax.tree.leaves(request.U_samples)[0].shape[0]
-    if batch_size == 1:
-        log_likelihoods = evaluate_one(
-            jax.tree.map(lambda values: values[0], request.U_samples)
-        )[None]
-    else:
-        log_likelihoods = jax.vmap(evaluate_one)(request.U_samples)
-    return LikelihoodEvaluation(log_likelihoods=log_likelihoods)
-
-
-@dataclasses.dataclass(slots=True, frozen=True)
-class ConstrainedSampleBatch(PureDataclassPytree):
-    """Fixed-width output of concurrent constrained-sampler lanes."""
-
-    U_samples: UType  # [S, ...] unit-hypercube pytree leaves
-    log_likelihoods: FloatArray  # [S]
-    num_likelihood_evaluations: IntArray  # [S]
-    phantom_samples: PhantomSamples  # [S, P, ...]
-    num_directions: IntArray  # [S]
-    num_isotropic: IntArray  # [S]
-
-
-ConstrainedSampleBatch.register_pytree()
-
-
-@dataclasses.dataclass(slots=True, frozen=True)
-class ConstrainedSampleRequest(PureDataclassPytree):
-    """Self-contained worker input for one constrained-sampling batch.
-
-    The model, arguments, parameters, and sampler configuration are registered
-    once per worker session. A request therefore carries only data that can
-    change between batches. Keeping seed coordinates in the request also means
-    an in-flight task never points back into a coordinator device buffer.
-    """
-
-    keys: PRNGKey  # [S, 2]
-    valid: BoolArray  # [S]
-    log_L_constraints: FloatArray  # [S]
-    seed_points: SeedPoint  # U0 [S, ...], log_L0 [S]
-    sampler_data: SamplerData | None  # [K, D, D] and aligned component data
-
-
-ConstrainedSampleRequest.register_pytree()
 
 
 def sample_request(
@@ -133,65 +83,6 @@ def sample_request(
         params=params,
     )
 
-
-def _sample_complete_chains(
-        sampler: AbstractSampler,
-        request: ConstrainedSampleRequest,
-        *,
-        args=(),
-        params=None,
-) -> ConstrainedSampleBatch:
-    """Run the reference complete-chain scalar or ``vmap`` implementation.
-
-    The local depth loop and worker program own the enclosing JIT boundary.
-    Keeping this compositional helper undecorated lets those boundaries capture
-    registered session objects such as notebook functions in ``args``; a
-    nested JIT would instead try to interpret them as dynamic arrays.
-    """
-
-    def sample_one(sample_key, constraint, seed_u, seed_log_likelihood):
-        seed = SeedPoint(U0=seed_u, log_L0=seed_log_likelihood)
-        return sampler.get_sample_with_diagnostics(
-            sample_key,
-            constraint,
-            seed,
-            args=args,
-            params=params,
-            sampler_data=request.sampler_data,
-        )
-
-    batch_size = request.log_L_constraints.shape[0]
-    if batch_size == 1:
-        sampled = sample_one(
-            request.keys[0],
-            request.log_L_constraints[0],
-            jax.tree.map(lambda values: values[0], request.seed_points.U0),
-            request.seed_points.log_L0[0],
-        )
-        sampled = jax.tree.map(lambda value: value[None], sampled)
-    else:
-        sampled = jax.vmap(sample_one)(
-            request.keys,
-            request.log_L_constraints,
-            request.seed_points.U0,
-            request.seed_points.log_L0,
-        )
-    (
-        U_samples,
-        log_likelihoods,
-        num_evals,
-        phantom_samples,
-        num_directions,
-        num_isotropic,
-    ) = sampled
-    return ConstrainedSampleBatch(
-        U_samples=U_samples,
-        log_likelihoods=log_likelihoods,
-        num_likelihood_evaluations=num_evals,
-        phantom_samples=phantom_samples,
-        num_directions=num_directions,
-        num_isotropic=num_isotropic,
-    )
 
 
 class AbstractSampler(ABC):
@@ -306,1039 +197,6 @@ class AbstractSampler(ABC):
 
 
 @partial(jax.jit, inline=True)
-def _sample_direction(key: PRNGKey, u0: TreeField[UType], radii: TreeField[UType] | None = None, rotation: UType | None = None) -> TreeField[UType]:
-    """
-    Choose a direction randomly from S^(D-1).
-
-    Args:
-        key: PRNG key
-        u0: a point in the sample space, used to determine the shape of the direction.
-
-    Returns:
-        direction: [D] direction from S^(D-1)
-    """
-    ndim = u0.ndim()
-    if ndim == 1:
-        return u0.ones_like()
-    direction = u0.random_normal_like(key)
-    if radii is not None:
-        direction = radii * direction
-    if rotation is not None:
-        direction = rotation @ direction
-    eps = jnp.asarray(1e-6, direction.norm().dtype)
-    norm = jnp.maximum(eps, direction.norm())
-    return direction / norm
-
-
-def _periodic_tree(
-        point: TreeField[UType],
-        periodic: tuple[bool, ...],
-) -> TreeField[UType]:
-    """Expand static scalar flags to the structure of one U-space point."""
-    leaves, structure = jax.tree.flatten(point.tree)
-    masks = []
-    offset = 0
-    for leaf in leaves:
-        size = int(np.prod(leaf.shape))
-        # [...] mask is constant in the compiled program and aligned to leaf.
-        mask = jnp.asarray(
-            periodic[offset:offset + size],
-            dtype=mp_policy.bool_dtype,
-        ).reshape(leaf.shape)
-        masks.append(mask)
-        offset += size
-    if offset != len(periodic):
-        raise ValueError(
-            "Periodic U-space topology does not match the sampled U shape."
-        )
-    return TreeField(jax.tree.unflatten(structure, masks))
-
-
-def _open_chart(
-        key: PRNGKey,
-        point: TreeField[UType],
-        periodic: tuple[bool, ...],
-) -> tuple[TreeField[UType], TreeField[UType]]:
-    """Translate periodic coordinates into an independent random chart."""
-    offset = point.random_uniform_like(key)
-    # The offset is state-independent. For a fixed offset this is a
-    # measure-preserving torus translation, so mixing over offsets preserves
-    # reversibility of the existing cube-bracket transition.
-    return _to_chart(point, offset, periodic), offset
-
-
-def _to_chart(
-        point: TreeField[UType],
-        offset: TreeField[UType],
-        periodic: tuple[bool, ...],
-) -> TreeField[UType]:
-    """Map a canonical point into an already selected random chart."""
-    mask = _periodic_tree(point, periodic)
-    return TreeField(jax.tree.map(
-        lambda value, shift, select: jnp.where(
-            select,
-            jnp.mod(value + shift, 1.0),
-            value,
-        ),
-        point.tree,
-        offset.tree,
-        mask.tree,
-    ))
-
-
-def _close_chart(
-        point: TreeField[UType],
-        offset: TreeField[UType],
-        periodic: tuple[bool, ...],
-) -> TreeField[UType]:
-    """Return a chart point to canonical U in the half-open unit cube."""
-    mask = _periodic_tree(point, periodic)
-    return TreeField(jax.tree.map(
-        lambda value, shift, select: jnp.where(
-            select,
-            jnp.mod(value - shift, 1.0),
-            value,
-        ),
-        point.tree,
-        offset.tree,
-        mask.tree,
-    ))
-
-
-def _slice_keys(
-        key: PRNGKey,
-        periodic: tuple[bool, ...],
-) -> tuple[PRNGKey, PRNGKey, PRNGKey, PRNGKey, PRNGKey]:
-    """Assign one independent chart key without perturbing the ordinary path."""
-    chart_key = key
-    if periodic:
-        chart_key, key = random.split(key, 2)
-    run_key, t_key, step_key, after_key = random.split(key, 4)
-    return chart_key, run_key, t_key, step_key, after_key
-
-
-def _sample_ellipsoidal_direction(
-        key: PRNGKey,
-        u0: TreeField[UType],
-        log_L_constraint: FloatArray,
-        data: SamplerData,
-        prob_isotropic: float,
-) -> TreeField[UType]:
-    """Draw from fixed contour-eligible geometry or its isotropic fallback."""
-    direction, _ = _draw_ellipsoidal_direction(
-        key,
-        u0,
-        log_L_constraint,
-        data,
-        prob_isotropic,
-    )
-    return direction
-
-
-def _draw_ellipsoidal_direction(
-        key: PRNGKey,
-        u0: TreeField[UType],
-        log_L_constraint: FloatArray,
-        data: SamplerData,
-        prob_isotropic: float,
-) -> tuple[TreeField[UType], BoolArray]:
-    """Draw a direction and report whether the isotropic kernel was used."""
-    isotropic_key, choice_key, component_key, normal_key = random.split(key, 4)
-    eligible = data.valid & (data.log_L_max > log_L_constraint)
-    use_isotropic = (
-        jnp.logical_not(jnp.any(eligible))
-        | (
-            random.uniform(choice_key, dtype=mp_policy.measure_dtype)
-            < jnp.asarray(prob_isotropic, mp_policy.measure_dtype)
-        )
-    )
-
-    def draw_ellipsoidal(_):
-        # Selection mass is geometric volume, not the EM mixture mass. The
-        # latter describes the weighted sample population and would bias this
-        # direction law towards densely sampled rather than spatially broad
-        # components.
-        logits = jnp.where(eligible, data.log_volumes, -jnp.inf)
-        component = random.categorical(component_key, logits).astype(
-            mp_policy.index_dtype
-        )
-        # The current values are intentionally ignored. Only their flattened
-        # structure and dtype define the direction space, keeping this law
-        # independent of the chain's current point.
-        flat, unravel = pytree_ravel(u0.tree)
-        normal = random.normal(normal_key, flat.shape, dtype=flat.dtype)
-        rotation = data.rotations[component].astype(flat.dtype)
-        radii = data.radii[component].astype(flat.dtype)
-        transformed = rotation @ (
-            radii * normal
-        )
-        norm = jnp.maximum(
-            jnp.linalg.norm(transformed),
-            jnp.asarray(1e-6, transformed.dtype),
-        )
-        return TreeField(unravel(transformed / norm))
-
-    direction = jax.lax.cond(
-        use_isotropic,
-        lambda unused: _sample_direction(isotropic_key, u0),
-        draw_ellipsoidal,
-        operand=None,
-    )
-    return direction, use_isotropic
-
-
-@partial(jax.jit, inline=True)
-def _slice_bounds(point_U0: TreeField[UType], direction: TreeField[UType]) -> tuple[FloatArray, FloatArray]:
-    """
-    Compute the slice bounds, t, where point_U0 + direction * t intersects uit cube boundary.
-
-    Args:
-        point_U0: starting point of slice
-        direction: direction of slice
-
-    Returns:
-        left_bound: left most point (<= 0).
-        right_bound: right most point (>= 0).
-    """
-    leaf_dtype = jax.tree.leaves(point_U0)[0].dtype
-    zero = jnp.zeros((), leaf_dtype)
-    one = jnp.ones((), leaf_dtype)
-    inf = jnp.full((), jnp.inf, leaf_dtype)
-    t1 = (one - point_U0) / direction
-    t1_right = jax.tree.reduce(lambda x, y: jnp.minimum(jnp.min(jnp.where(x >= zero, x, inf)), jnp.min(jnp.where(y >= zero, y, inf))), t1,
-                               initializer=jnp.inf)
-    t1_left = jax.tree.reduce(lambda x, y: jnp.maximum(jnp.max(jnp.where(x <= zero, x, -inf)), jnp.max(jnp.where(y <= zero, y, -inf))), t1,
-                              initializer=-jnp.inf)
-    t0 = -point_U0 / direction
-    t0_right = jax.tree.reduce(lambda x, y: jnp.minimum(jnp.min(jnp.where(x >= zero, x, inf)), jnp.min(jnp.where(y >= zero, y, inf))), t0,
-                               initializer=jnp.inf)
-    t0_left = jax.tree.reduce(lambda x, y: jnp.maximum(jnp.max(jnp.where(x <= zero, x, -inf)), jnp.max(jnp.where(y <= zero, y, -inf))), t0,
-                              initializer=-jnp.inf)
-    right_bound = jnp.minimum(t0_right, t1_right)
-    left_bound = jnp.maximum(t0_left, t1_left)
-    return left_bound, right_bound
-
-
-@partial(jax.jit, inline=True)
-def _pick_point_in_interval(key: PRNGKey, point_U0: TreeField[UType], direction: TreeField[UType], left: FloatArray,
-                            right: FloatArray) -> tuple[TreeField[UType], FloatArray]:
-    """
-    Select a point along slice in [point_U0 + direction * left, point_U0 + direction * right]
-
-    Args:
-        key: PRNG key
-        point_U0: [D]
-        direction: [D]
-        left: left most point (<= 0).
-        right: right most point (>= 0).
-
-    Returns:
-        point_U: [D]
-        t: selection point between [left, right]
-    """
-    leaf_dtype = jax.tree.leaves(point_U0)[0].dtype
-    t = left + random.uniform(key, dtype=leaf_dtype) * (right - left)
-    point_U = point_U0 + direction * t
-    return point_U, t
-
-
-@partial(jax.jit, inline=True)
-def _shrink_interval(t: FloatArray, left: FloatArray, right: FloatArray) -> tuple[FloatArray, FloatArray]:
-    """
-    Not successful proposal, so shrink, optionally apply exponential shrinkage.
-    """
-    zero = jnp.zeros_like(t)
-    left = jnp.where(t < zero, t, left)
-    right = jnp.where(t > zero, t, right)
-    return left, right
-
-
-def _new_proposal(
-        key: PRNGKey,
-        U0: TreeField[UType],
-        direction: TreeField[UType],
-        slice_width: FloatArray,
-        no_step_out: bool,
-        gradient_guided: bool,
-        log_L_constraint: FloatArray,
-        log_likelihood_fn: Callable[[UType], FloatArray],
-        periodic: tuple[bool, ...] = (),
-        sampler_data: SamplerData | None = None,
-        prob_isotropic: float = 1.0,
-) -> tuple[
-    TreeField[UType],
-    FloatArray,
-    IntArray,
-    TreeField[UType],
-    FloatArray,
-    BoolArray,
-]:
-    """
-    Sample from a slice about a seed point.
-
-    Args:
-        key: PRNG key
-        direction: the direction to sample along
-        no_step_out: if true then perform exponential shrinkage from maximal bounds, requiring no step-out procedure.
-        gradient_guided: if true then do householder reflections
-        log_L_constraint: the constraint to sample within
-        log_likelihood_fn: the log-likelihood function
-
-    Returns:
-        point_U: the new sample
-        log_L: the log-likelihood of the new sample
-        num_likelihood_evaluations: the number of likelihood evaluations performed
-    """
-
-    class Carry(NamedTuple):
-        key: PRNGKey
-        direction: TreeField[UType]
-        left: FloatArray
-        right: FloatArray
-        t: FloatArray
-        point_U: TreeField[UType]
-        log_L: FloatArray
-        num_likelihood_evaluations: IntArray
-
-    def cond(carry: Carry) -> BoolArray:
-        satisfaction = carry.log_L > log_L_constraint
-        return jnp.bitwise_not(satisfaction)
-
-    def body(carry: Carry) -> Carry:
-        key, t_key = random.split(carry.key, 2)
-        left, right = _shrink_interval(
-            t=carry.t,
-            left=carry.left,
-            right=carry.right,
-        )
-        point_U, t = _pick_point_in_interval(
-            key=t_key,
-            point_U0=U0,
-            direction=carry.direction,
-            left=left,
-            right=right
-        )
-        log_L = log_likelihood_fn(point_U.tree).astype(log_L_constraint.dtype)
-        num_likelihood_evaluations = carry.num_likelihood_evaluations + jnp.ones_like(carry.num_likelihood_evaluations)
-        return Carry(
-            key=key,
-            t=t,
-            left=left,
-            right=right,
-            point_U=point_U,
-            log_L=log_L,
-            num_likelihood_evaluations=num_likelihood_evaluations,
-            direction=carry.direction
-        )
-
-    # Chose the direction to go
-    num_likelihood_evaluations = jnp.full((), 0, mp_policy.count_dtype)
-
-    chart_key, run_key, t_key, step_key, after_key = _slice_keys(
-        key,
-        periodic,
-    )
-    chart_offset = None
-    if periodic:
-        # Perfect bracketing remains finite because it runs in a randomly cut
-        # cube chart. Likelihoods and retained samples always use canonical U.
-        U0, chart_offset = _open_chart(chart_key, U0, periodic)
-        canonical_log_likelihood_fn = log_likelihood_fn
-
-        def log_likelihood_fn(value):
-            canonical = _close_chart(
-                TreeField(value),
-                chart_offset,
-                periodic,
-            )
-            return canonical_log_likelihood_fn(canonical.tree)
-
-    (left_bound, right_bound) = _slice_bounds(
-        point_U0=U0,
-        direction=direction
-    )
-    slice_width = jnp.asarray(slice_width, left_bound.dtype)
-
-    if no_step_out:
-        left, right = left_bound, right_bound
-        step_out_num_likelihood_evaluations = jnp.asarray(0, mp_policy.count_dtype)
-    else:
-        class StepOutCarry(NamedTuple):
-            key: PRNGKey
-            left: FloatArray
-            right: FloatArray
-            left_outside_slice: BoolArray
-            right_outside_slice: BoolArray
-            num_likelihood_evaluations: IntArray
-
-        eps = jnp.asarray(1e-12, left_bound.dtype)
-        use_full_slice = jnp.isinf(slice_width)
-        effective_slice_width = jnp.maximum(slice_width, eps)
-        place_key, step_key = random.split(step_key)
-        uniform_origin = random.uniform(place_key, dtype=left_bound.dtype)
-        initial_left = -uniform_origin * effective_slice_width
-        initial_right = initial_left + effective_slice_width
-
-        left = jnp.where(use_full_slice, left_bound, jnp.maximum(left_bound, initial_left))
-        right = jnp.where(use_full_slice, right_bound, jnp.minimum(right_bound, initial_right))
-
-        def _point_at_t(t: FloatArray) -> TreeField[UType]:
-            return U0 + direction * t
-
-        def step_out_cond(carry: StepOutCarry) -> BoolArray:
-            can_expand_left = carry.left > left_bound
-            can_expand_right = carry.right < right_bound
-            both_outside_box = jnp.bitwise_not(jnp.bitwise_or(can_expand_left, can_expand_right))
-            both_outside_slice = jnp.bitwise_and(carry.left_outside_slice, carry.right_outside_slice)
-            return jnp.bitwise_not(jnp.bitwise_or(both_outside_box, both_outside_slice))
-
-        def step_out_body(carry: StepOutCarry) -> StepOutCarry:
-            key, choose_key = random.split(carry.key, 2)
-
-            can_expand_left = carry.left > left_bound
-            can_expand_right = carry.right < right_bound
-
-            choose_left_random = random.uniform(choose_key, dtype=left_bound.dtype) < jnp.where(
-                can_expand_left & can_expand_right, 0.5, jnp.where(can_expand_left, 1., 0.))
-
-            current_width = jnp.maximum(carry.right - carry.left, eps)
-            candidate_left = jnp.maximum(left_bound, carry.left - current_width)
-            candidate_right = jnp.minimum(right_bound, carry.right + current_width)
-
-            t_eval = jnp.where(choose_left_random, candidate_left, candidate_right)
-
-            next_left = jnp.where(choose_left_random, candidate_left, carry.left)
-            next_right = jnp.where(choose_left_random, carry.right, candidate_right)
-
-            log_L_eval = log_likelihood_fn(_point_at_t(t_eval).tree)
-            outside_slice_eval = log_L_eval <= log_L_constraint
-            next_left_outside_slice = jnp.where(choose_left_random, outside_slice_eval, carry.left_outside_slice)
-            next_right_outside_slice = jnp.where(choose_left_random, carry.right_outside_slice, outside_slice_eval)
-            num_likelihood_evaluations = carry.num_likelihood_evaluations + jnp.ones_like(carry.num_likelihood_evaluations)
-
-            return StepOutCarry(
-                key=key,
-                left=next_left,
-                right=next_right,
-                left_outside_slice=next_left_outside_slice,
-                right_outside_slice=next_right_outside_slice,
-                num_likelihood_evaluations=num_likelihood_evaluations
-            )
-
-        left_outside_slice = log_likelihood_fn(_point_at_t(left).tree) <= log_L_constraint
-        right_outside_slice = log_likelihood_fn(_point_at_t(right).tree) <= log_L_constraint
-        step_out_init = StepOutCarry(
-            key=step_key,
-            left=left,
-            right=right,
-            left_outside_slice=left_outside_slice,
-            right_outside_slice=right_outside_slice,
-            num_likelihood_evaluations=jnp.asarray(2, mp_policy.count_dtype)
-        )
-
-        step_out_carry = jax.lax.while_loop(
-            cond_fun=lambda c: jnp.bitwise_and(jnp.bitwise_not(use_full_slice), step_out_cond(c)),
-            body_fun=step_out_body,
-            init_val=step_out_init
-        )
-        left, right = step_out_carry.left, step_out_carry.right
-        step_out_num_likelihood_evaluations = step_out_carry.num_likelihood_evaluations
-
-    point_U, t = _pick_point_in_interval(
-        key=t_key,
-        point_U0=U0,
-        direction=direction,
-        left=left,
-        right=right
-    )
-    log_L = log_likelihood_fn(point_U.tree).astype(log_L_constraint.dtype)
-    num_likelihood_evaluations += jnp.ones_like(num_likelihood_evaluations)
-    num_likelihood_evaluations += step_out_num_likelihood_evaluations
-    init_carry = Carry(
-        key=run_key,
-        direction=direction,
-        left=left,
-        right=right,
-        t=t,
-        point_U=point_U,
-        log_L=log_L,
-        num_likelihood_evaluations=num_likelihood_evaluations
-    )
-
-    carry = jax.lax.while_loop(
-        cond_fun=cond,
-        body_fun=body,
-        init_val=init_carry
-    )
-
-    # Update direction
-    direction = carry.direction
-    num_likelihood_evaluations = carry.num_likelihood_evaluations
-    if gradient_guided:
-        # Perform HMC with Householder reflections
-        raise NotImplementedError("Gradient guided slice sampler not implemented.")
-    else:
-        # Randomly choose a new direction
-        if sampler_data is None:
-            direction = _sample_direction(after_key, direction)
-            direction_isotropic = jnp.asarray(True, mp_policy.bool_dtype)
-        else:
-            direction, direction_isotropic = _draw_ellipsoidal_direction(
-                after_key,
-                direction,
-                log_L_constraint,
-                sampler_data,
-                prob_isotropic,
-            )
-    next_slice_width = 2 * (carry.right - carry.left)
-    point_U = carry.point_U
-    if periodic:
-        point_U = _close_chart(point_U, chart_offset, periodic)
-    return (
-        point_U,
-        carry.log_L,
-        num_likelihood_evaluations,
-        direction,
-        next_slice_width,
-        direction_isotropic,
-    )
-
-
-@dataclasses.dataclass(slots=True, frozen=True)
-class SliceBatchState(PureDataclassPytree):
-    """Logical slice chains continued between batched likelihood calls."""
-
-    anchor: UType  # [S, ...] last accepted point in each chain
-    chart_offset: UType | None  # [S, ...] periodic chart shift; otherwise None
-    log_likelihood: FloatArray  # [S]
-    directions: UType  # [S, T, ...] direction for each transition
-    direction_is_isotropic: BoolArray  # [S, T]
-    left: FloatArray  # [S]
-    right: FloatArray  # [S]
-    t: FloatArray  # [S]
-    proposal: UType  # [S, ...] point awaiting likelihood evaluation
-    run_key: PRNGKey  # [S, 2]
-    transition_keys: PRNGKey  # [S, T, 2]
-    transition_index: IntArray  # [S]
-    done: BoolArray  # [S]
-    num_likelihood_evaluations: IntArray  # [S] logical calls only
-    phantom_samples: UType  # [S, P, ...] retained start prefix
-    phantom_log_likelihoods: FloatArray  # [S, P]
-
-
-SliceBatchState.register_pytree()
-
-
-def _initialise_slice_chains(
-        sampler: UniDimSliceSampler,
-        request: ConstrainedSampleRequest,
-) -> SliceBatchState:
-    """Prepare one unevaluated proposal per logical slice chain."""
-    num_slices = sampler.num_slices
-    num_phantom = sampler.num_phantom()
-    sampler_data = request.sampler_data
-
-    def initialise_one(
-            key,
-            valid,
-            log_L_constraint,
-            seed_u,
-            seed_log_likelihood,
-    ):
-        def draw_direction(direction_key):
-            if sampler_data is None:
-                return (
-                    _sample_direction(direction_key, TreeField(seed_u)),
-                    jnp.asarray(True, mp_policy.bool_dtype),
-                )
-            return _draw_ellipsoidal_direction(
-                direction_key,
-                TreeField(seed_u),
-                log_L_constraint,
-                sampler_data,
-                sampler.direction.prob_isotropic,
-            )
-
-        # Preserve the scalar sampler's random-key schedule exactly. Pool
-        # execution order must never change a logical chain's random stream.
-        direction_key, sample_key = random.split(key, 2)
-        initial_direction, initial_is_isotropic = draw_direction(
-            direction_key
-        )
-        sample_key, initial_key = random.split(sample_key, 2)
-        later_keys = random.split(sample_key, num_slices - 1)
-        transition_keys = jnp.concatenate(
-            [initial_key[None], later_keys],
-            axis=0,
-        )
-
-        # Direction proposals depend on the fixed parent contour and unit-cube
-        # structure, not on accepted chain coordinates. Hoisting this stream
-        # avoids evaluating both direction branches inside a vmapped condition
-        # for every rejected proposal, while preserving the scalar direction
-        # law and the requirement that every transition uses the same law.
-        # The compile, memory, and execution trade-off is recorded in
-        # benchmarks/issue_244/REPORT.md.
-        if num_slices == 1:
-            directions = jax.tree.map(
-                lambda value: value[None],
-                initial_direction.tree,
-            )
-            directions_are_isotropic = initial_is_isotropic[None]
-        else:
-            def draw_later_direction(transition_key):
-                _, _, _, _, after_key = _slice_keys(
-                    transition_key,
-                    sampler._periodic,
-                )
-                return draw_direction(after_key)
-
-            # The scalar reference draws one direction per scan step. Keep
-            # that transition ordering so XLA performs the same matrix-vector
-            # reductions for ellipsoidal directions; batching the transition
-            # axis itself can change floating-point reduction order enough to
-            # send a difficult chain down a different slice path.
-            def scan_direction(unused, transition_key):
-                return unused, draw_later_direction(transition_key)
-
-            _, (
-                later_directions,
-                later_is_isotropic,
-            ) = jax.lax.scan(
-                scan_direction,
-                None,
-                transition_keys[:-1],
-            )
-            directions = jax.tree.map(
-                lambda initial, later: jnp.concatenate(
-                    [initial[None], later],
-                    axis=0,
-                ),
-                initial_direction.tree,
-                later_directions.tree,
-            )
-            directions_are_isotropic = jnp.concatenate(
-                [initial_is_isotropic[None], later_is_isotropic],
-                axis=0,
-            )
-
-        # Perfect bracketing always has exactly one proposal ready. A rejected
-        # likelihood shrinks this interval; an accepted likelihood advances
-        # this logical chain to its next slice transition without a barrier.
-        chart_key, run_key, t_key, _, _ = _slice_keys(
-            transition_keys[0],
-            sampler._periodic,
-        )
-        chart_offset = None
-        chart_seed = TreeField(seed_u)
-        if sampler._periodic:
-            chart_seed, chart_offset = _open_chart(
-                chart_key,
-                chart_seed,
-                sampler._periodic,
-            )
-        left, right = _slice_bounds(
-            chart_seed,
-            initial_direction,
-        )
-        proposal, t = _pick_point_in_interval(
-            t_key,
-            chart_seed,
-            initial_direction,
-            left,
-            right,
-        )
-        if sampler._periodic:
-            proposal = _close_chart(
-                proposal,
-                chart_offset,
-                sampler._periodic,
-            )
-        phantom_samples = jax.tree.map(
-            lambda value: jnp.zeros(
-                (num_phantom,) + value.shape,
-                value.dtype,
-            ),
-            seed_u,
-        )
-        return SliceBatchState(
-            anchor=seed_u,
-            chart_offset=(
-                None if chart_offset is None else chart_offset.tree
-            ),
-            log_likelihood=jnp.asarray(
-                seed_log_likelihood,
-                log_L_constraint.dtype,
-            ),
-            directions=directions,
-            direction_is_isotropic=directions_are_isotropic,
-            left=left,
-            right=right,
-            t=t,
-            proposal=proposal.tree,
-            run_key=run_key,
-            transition_keys=transition_keys,
-            transition_index=jnp.asarray(0, mp_policy.index_dtype),
-            # Static scheduler padding is not a logical chain. Mark it done
-            # immediately so an unused lane cannot extend the physical loop;
-            # it still occupies the fixed likelihood width as neutral filler
-            # while any valid chain remains active.
-            done=jnp.bitwise_not(valid),
-            num_likelihood_evaluations=jnp.asarray(
-                0,
-                mp_policy.count_dtype,
-            ),
-            phantom_samples=phantom_samples,
-            phantom_log_likelihoods=jnp.full(
-                (num_phantom,),
-                -jnp.inf,
-                log_L_constraint.dtype,
-            ),
-        )
-
-    return jax.vmap(initialise_one)(
-        request.keys,
-        request.valid,
-        request.log_L_constraints,
-        request.seed_points.U0,
-        request.seed_points.log_L0,
-    )
-
-
-def _continue_slice_chains(
-        sampler: UniDimSliceSampler,
-        request: ConstrainedSampleRequest,
-        *,
-        args=(),
-        params=None,
-) -> ConstrainedSampleBatch:
-    """Batch ready likelihoods while each logical chain advances freely.
-
-    The physical likelihood width equals the logical request width. A finished
-    lane evaluates the neutral unit-cube point ``U=0.5`` until the slowest
-    complete chain finishes. These filler calls are substantially fewer than
-    the masked calls introduced by a barrier after every slice transition,
-    while the reported counter remains the exact logical count.
-
-    As with the complete-chain reference, the enclosing local or worker
-    program owns JIT compilation so registered non-array session objects remain
-    captured Python closure values.
-    """
-    num_chains = request.log_L_constraints.shape[0]
-    num_slices = sampler.num_slices
-    num_phantom = sampler.num_phantom()
-    assert 0 <= num_phantom <= num_slices - 1, (
-        "num_phantom() should be in [0, num_slices - 1]"
-    )
-
-    def evaluate_one(u_sample):
-        return sampler.model.log_likelihood(
-            u_sample,
-            args=args,
-            params=params,
-            allow_nan=False,
-        ).astype(request.log_L_constraints.dtype)
-
-    def prepare_transition(
-            chain,
-            anchor,
-            direction,
-            transition_index,
-    ):
-        transition_key = chain.transition_keys[transition_index]
-        chart_key, run_key, t_key, _, _ = _slice_keys(
-            transition_key,
-            sampler._periodic,
-        )
-        chart_offset = None
-        chart_anchor = TreeField(anchor)
-        if sampler._periodic:
-            chart_anchor, chart_offset = _open_chart(
-                chart_key,
-                chart_anchor,
-                sampler._periodic,
-            )
-        left, right = _slice_bounds(
-            chart_anchor,
-            direction,
-        )
-        proposal, t = _pick_point_in_interval(
-            t_key,
-            chart_anchor,
-            direction,
-            left,
-            right,
-        )
-        if sampler._periodic:
-            proposal = _close_chart(
-                proposal,
-                chart_offset,
-                sampler._periodic,
-            )
-        return dataclasses.replace(
-            chain,
-            anchor=anchor,
-            chart_offset=(
-                None if chart_offset is None else chart_offset.tree
-            ),
-            left=left,
-            right=right,
-            t=t,
-            proposal=proposal.tree,
-            run_key=run_key,
-        )
-
-    def consume_one(chain, log_likelihood, active, constraint):
-        def consume(active_chain):
-            transition_index = active_chain.transition_index
-            active_chain = dataclasses.replace(
-                active_chain,
-                log_likelihood=log_likelihood,
-                num_likelihood_evaluations=(
-                    active_chain.num_likelihood_evaluations
-                    + jnp.asarray(1, mp_policy.count_dtype)
-                ),
-            )
-
-            def accept(accepted_chain):
-                if num_phantom == 0:
-                    phantom_samples = accepted_chain.phantom_samples
-                    phantom_log_likelihoods = (
-                        accepted_chain.phantom_log_likelihoods
-                    )
-                else:
-                    phantom_index = jnp.minimum(
-                        transition_index,
-                        num_phantom - 1,
-                    )
-                    retain = transition_index < num_phantom
-                    phantom_samples = jax.tree.map(
-                        lambda history, value: history.at[
-                            phantom_index
-                        ].set(
-                            jnp.where(
-                                retain,
-                                value,
-                                history[phantom_index],
-                            )
-                        ),
-                        accepted_chain.phantom_samples,
-                        accepted_chain.proposal,
-                    )
-                    phantom_log_likelihoods = (
-                        accepted_chain.phantom_log_likelihoods.at[
-                            phantom_index
-                        ].set(
-                            jnp.where(
-                                retain,
-                                log_likelihood,
-                                accepted_chain.phantom_log_likelihoods[
-                                    phantom_index
-                                ],
-                            )
-                        )
-                    )
-                next_index = transition_index + jnp.asarray(
-                    1,
-                    mp_policy.index_dtype,
-                )
-                finished = next_index == num_slices
-                accepted_chain = dataclasses.replace(
-                    accepted_chain,
-                    anchor=accepted_chain.proposal,
-                    phantom_samples=phantom_samples,
-                    phantom_log_likelihoods=phantom_log_likelihoods,
-                    transition_index=next_index,
-                    done=finished,
-                )
-
-                def prepare_next(unfinished_chain):
-                    # A batched ``cond`` traces this branch for a finished lane
-                    # too. The safe index makes that unused value explicit and
-                    # avoids relying on out-of-range gather clamping.
-                    safe_index = jnp.minimum(next_index, num_slices - 1)
-                    direction = TreeField(
-                        jax.tree.map(
-                            lambda values: values[safe_index],
-                            unfinished_chain.directions,
-                        )
-                    )
-                    return prepare_transition(
-                        unfinished_chain,
-                        unfinished_chain.anchor,
-                        direction,
-                        safe_index,
-                    )
-
-                return jax.lax.cond(
-                    finished,
-                    lambda value: value,
-                    prepare_next,
-                    accepted_chain,
-                )
-
-            def reject(rejected_chain):
-                run_key, t_key = random.split(
-                    rejected_chain.run_key,
-                    2,
-                )
-                left, right = _shrink_interval(
-                    rejected_chain.t,
-                    rejected_chain.left,
-                    rejected_chain.right,
-                )
-                direction = TreeField(
-                    jax.tree.map(
-                        lambda values: values[transition_index],
-                        rejected_chain.directions,
-                    )
-                )
-                chart_anchor = TreeField(rejected_chain.anchor)
-                if sampler._periodic:
-                    chart_offset = TreeField(
-                        rejected_chain.chart_offset
-                    )
-                    # Retrying a rejected proposal must retain the transition's
-                    # existing chart; drawing another seam here would change
-                    # the slice interval and invalidate shrinkage.
-                    chart_anchor = _to_chart(
-                        chart_anchor,
-                        chart_offset,
-                        sampler._periodic,
-                    )
-                proposal, t = _pick_point_in_interval(
-                    t_key,
-                    chart_anchor,
-                    direction,
-                    left,
-                    right,
-                )
-                if sampler._periodic:
-                    proposal = _close_chart(
-                        proposal,
-                        chart_offset,
-                        sampler._periodic,
-                    )
-                return dataclasses.replace(
-                    rejected_chain,
-                    left=left,
-                    right=right,
-                    t=t,
-                    proposal=proposal.tree,
-                    run_key=run_key,
-                )
-
-            return jax.lax.cond(
-                log_likelihood > constraint,
-                accept,
-                reject,
-                active_chain,
-            )
-
-        return jax.lax.cond(
-            active,
-            consume,
-            lambda value: value,
-            chain,
-        )
-
-    def continue_condition(state):
-        return jnp.any(jnp.bitwise_not(state.done))
-
-    def continue_body(state):
-        active = jnp.bitwise_not(state.done)
-        # Static-width filler keeps one likelihood program and lets scientific
-        # likelihoods retain ordinary vmap semantics. Filler results never
-        # enter chain state or the user-visible logical evaluation count.
-        proposals = jax.tree.map(
-            lambda values: jnp.where(
-                active.reshape(
-                    (num_chains,) + (1,) * (values.ndim - 1)
-                ),
-                values,
-                jnp.full_like(values, 0.5),
-            ),
-            state.proposal,
-        )
-        if num_chains == 1:
-            log_likelihoods = evaluate_one(
-                jax.tree.map(lambda values: values[0], proposals)
-            )[None]
-        else:
-            log_likelihoods = jax.vmap(evaluate_one)(proposals)
-        return jax.vmap(consume_one)(
-            state,
-            log_likelihoods,
-            active,
-            request.log_L_constraints,
-        )
-
-    state = jax.lax.while_loop(
-        continue_condition,
-        continue_body,
-        _initialise_slice_chains(sampler, request),
-    )
-    phantom_samples = PhantomSamples(
-        U_samples=state.phantom_samples,
-        log_L=state.phantom_log_likelihoods,
-        valid_mask=jnp.broadcast_to(
-            request.valid[:, None],
-            (num_chains, num_phantom),
-        ),
-    )
-    if sampler.direction is None:
-        num_directions = jnp.zeros(
-            (num_chains,),
-            mp_policy.count_dtype,
-        )
-        num_isotropic = jnp.zeros(
-            (num_chains,),
-            mp_policy.count_dtype,
-        )
-    else:
-        num_directions = jnp.full(
-            (num_chains,),
-            num_slices,
-            mp_policy.count_dtype,
-        )
-        num_isotropic = jnp.sum(
-            state.direction_is_isotropic.astype(mp_policy.count_dtype),
-            axis=1,
-        )
-    return ConstrainedSampleBatch(
-        U_samples=state.anchor,
-        log_likelihoods=state.log_likelihood,
-        num_likelihood_evaluations=state.num_likelihood_evaluations,
-        phantom_samples=phantom_samples,
-        num_directions=num_directions,
-        num_isotropic=num_isotropic,
-    )
-
-
-@partial(jax.jit, inline=True)
-def get_seed_point(key: PRNGKey, samples: Samples, log_L_constraint: FloatArray) -> SeedPoint:
-    """
-    Get a seed point from samples that satisfies the likelihood constraint. This is done by masking samples that do not satisfy the constraint, and sampling uniformly from the remaining samples.
-
-    Args:
-        key: PRNGKey
-        samples: samples to select from.
-        log_L_constraint: the constraint to sample within.
-
-    Returns:
-        seed point that satisfy the constraint.
-    """
-    select_mask = samples.log_likelihoods > log_L_constraint
-    seed_point = sample_uniformly_masked(
-        key=key,
-        v=SeedPoint(U0=samples.U_samples, log_L0=samples.log_likelihoods),
-        select_mask=select_mask,
-        num_samples=1,
-        squeeze=True
-    )
-    return seed_point
-
-
 def _take_phantom_prefix(cumulative_samples, num_phantom: int):
     """Take retained generated transitions from the start of a chain."""
     return jax.tree.map(lambda x: x[:num_phantom], cumulative_samples)
